refactor : 코드 분할 및 리팩토링
