+++ src/App.js
@@ -3,11 +3,13 @@ import { PROMPTS } from './lib/prompt.js';
 class App {
   async run() {
     const controller = new Controller();
-    controller.outputView.printMessage(PROMPTS.intro);
+
+    controller.intro();
     await controller.inputDate();
     await controller.inputMenu();
-    controller.outputView.printMessage(PROMPTS.preview);
-    controller.run();
+
+    const resultObj = controller.evaluateResult();
+    controller.printResult(resultObj);
   }
 }
